backend/requirements.txt
Flask
bcrypt
PyJWT
requests